Add files via upload
This file contains some functions that visualises the buildings in Creswell, North Carolina. Firstly, we have a map with all the buildings in the town highlighted in red along with the terrain of the area. Secondly, we have a function that takes in the GeoDataFrame and plots a histogram for one of its attributes, in this case the monetary value of said buildings, and then also fits a normal distribution curve over the plot. Lastly, we can use that distribution create a simulated set of data for future model usage.